embed/requirements.txt
pinecone-client
langchain
tiktoken
unstructured
openai
python-dotenv
langchain_openai
"unstructured[md]"